refactor: #668 package 오타 수정
