+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/data/di/LocalDatabaseModule.kt
@@ -1,8 +1,8 @@
 package com.team.todoktodok.data.di
 
 import android.content.Context
-import com.team.todoktodok.data.local.nofitication.DefaultNotificationDataStore
-import com.team.todoktodok.data.local.nofitication.NotificationDataStore
+import com.team.todoktodok.data.local.notification.DefaultNotificationDataStore
+import com.team.todoktodok.data.local.notification.NotificationDataStore
 
 class LocalDatabaseModule(
     context: Context,
